adds shorter tests to make testing quicker

+++ lib/item.rb
@@ -14,7 +14,7 @@ class Item
     @id = hash[:id].to_i
     @name = hash[:name]
     @description = hash[:description]
-    @unit_price = BigDecimal.new(hash[:unit_price].to_d/100)
+    @unit_price = BigDecimal.new(hash[:unit_price].to_i/100)
     @created_at = hash[:created_at]
     @updated_at = hash[:updated_at]
     @merchant_id = hash[:merchant_id].to_i
